Modify the Word document
Version Management
It is the process of keeping track of different versions of software components or configuration items and the systems in which these components are used. 
It also involves ensuring that changes made by different developers to these versions do not interfere with each other. 
It can also be thought of as the process of managing codelines and baselines. 
Codelines normally apply to components of systems so that there are different versions of each component.